Fix #104 (path pollution)
_in_process.py is executed as a subprocess. This causes its parent
directory to become sys.path[0]. Any extra files in the same directory
shadow pypi libraries eg colorlog.

Solution: put _in_process.py in its own special subdirectory.
Note: the directory has to be a module because
importlib.resources.path does not support resources with directory
separators in their name:

https://docs.python.org/3/library/importlib.html#importlib.resources.path :
> resource is the name of the resource to open within package;
> it may not contain path separators and it may not have sub-resources
> (i.e. it cannot be a directory).
